Clean up main viewmodel
