+++ config/detekt/detekt.yml
@@ -138,7 +138,7 @@ complexity:
     ignoreDefaultParameters: false
     ignoreDataClasses: true
     ignoreAnnotatedParameter: []
-    ignoreAnnotated: ['Composable']
+    ignoreAnnotated: ['Composable', 'HiltViewModel']
   MethodOverloading:
     active: false
     threshold: 6
